+++ src/main/java/com/backendoori/ootw/controller/PostController.java
@@ -21,6 +21,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import service.PostService;
 
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/posts")
