+++ src/main/java/service/PostService.java
@@ -1,4 +1,4 @@
-package com.backendoori.ootw;
+package service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
