+++ src/test/java/com/backendoori/ootw/PostServiceTest.java
@@ -23,6 +23,7 @@ import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import service.PostService;
 
 @SpringBootTest
 class PostServiceTest {
